수정

+++ src/main/java/com/moongcheap_backend/payout/infrastructure/PayoutRepository.java
@@ -3,6 +3,6 @@ package com.moongcheap_backend.payout.infrastructure;
 import com.moongcheap_backend.payout.domain.Payout;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PayoutRepository extends JpaRepository<Payout, Integer> {
+public interface PayoutRepository extends JpaRepository<Payout, Long> {
 
 }
